requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
packaging==23.0
pdf2image==1.16.2
Pillow==9.4.0
pytesseract==0.3.10
requests==2.28.2
urllib3==1.26.14
